Text To Voice Converter
In this we will learn to make a Text to speech converter or Text to voice generator app using HTML CSS and JavaScript. We will also add feature to change the voice. We can generate the speech in different voices.
This is best JavaScript project for college student or job portfolio.

Download Images: https://drive.google.com/file/d/1dvOKzOZ-ENcUgnI1q3FDiOsCoZBGXVx1/view